Converting to 32 bits I part The default I part must be set to default<<8, if default was the value using 16 bits.

+++ services/heating_controller/heating_ctrl.h
@@ -38,7 +38,7 @@ typedef struct
   int16_t uMax;
 
   // States
-  int16_t I;
+  int32_t I;
   int16_t u;
 
 } pid_data_t;
